fix(pipeline): lockfile commonjs
