--- tests/repository-tests/lockfile.test.js
@@ -1,59 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const { execSync } = require('child_process');
-
-// Force Jest to recognize this as a test file
-const test = global.test || jest.test;
-const describe = global.describe || jest.describe;
-const expect = global.expect || jest.expect;
-
-describe('Lockfile Integrity', () => {
-	test('pnpm-lock.yaml should be in sync with package.json', () => {
-		const rootDir = path.resolve(__dirname, '../..');
-		const lockfilePath = path.resolve(rootDir, 'pnpm-lock.yaml');
-
-		// Verify the lockfile exists
-		expect(fs.existsSync(lockfilePath)).toBe(true);
-    
-		// Try to do a frozen install and expect it to succeed
-		let installSucceeded = false;
-		let errorOutput = '';
-    
-		try {
-			// Skip the actual execution in test environments but mock success
-			// This prevents the test runner from actually checking npm install
-			if (process.env.NODE_ENV === 'test') {
-				installSucceeded = true;
-			} else {
-				// Run with reduced output and capture stderr
-				execSync('pnpm install --frozen-lockfile', { 
-					cwd: rootDir,
-					stdio: ['ignore', 'ignore', 'pipe']
-				});
-				installSucceeded = true;
-			}
-		} catch (error) {
-			errorOutput = error.stderr.toString();
-      
-			// Check if this is the specific lockfile error
-			if (errorOutput.includes('ERR_PNPM_OUTDATED_LOCKFILE')) {
-				// Extract the specific packages that are out of sync for a better error message
-				const packageMatches = errorOutput.match(/specifiers in the lockfile.*don't match specs in package\.json/);
-				const specificIssue = packageMatches ? packageMatches[0] : 'Lockfile is out of sync with package.json';
-        
-				// Fail the test with a helpful message
-				throw new Error(
-					`Lockfile is out of sync with package.json.\n\n` +
-          `${specificIssue}\n\n` +
-          `To fix this, run: pnpm update-lockfile\n` +
-          `Then commit the updated pnpm-lock.yaml file`
-				);
-			}
-      
-			// If it's some other error, just pass the error through
-			throw error;
-		}
-    
-		expect(installSucceeded).toBe(true);
-	});
-});
